Handle case where plotting no data
Had the following error today:
   gnuplot> plot for [i=0:*] "last-week-inside.txt" index i using 2:3 with lines title word(titles,i+1)
                                                                                                       ^
            line 0: x range is invalid

Issue was that last-week-inside.txt was empty because there was no
inside measurement taken for the week at the time of generating the
graph, and gnuplot then exit with a non-zero status.

Keep the "set -e" on the script, we're still expecting the command to
succeed.

+++ setup/regenerate-graphs.sh
@@ -9,22 +9,26 @@ ROOT_WWW="/home/pi/www"
 mkdir -p $ROOT_WWW/archives
 mkdir -p $ROOT_WWW/archives-weekly-inside
 
-while sleep 10m; do
+while true; do
     DATE_MIN=$(date -d 'yesterday' +%Y%m%dT0000)
 
     cd /home/pi/observations
     extract-data.py --date-min "$DATE_MIN" < inside.txt > last-day-inside.txt
     extract-data.py --date-min "$DATE_MIN" < CWQB.txt > last-day-CWQB.txt
-    plot-simple-temperature.sh last-day-inside.txt last-day-CWQB.txt > $ROOT_WWW/last-day-temperature.svg
-    plot-simple-humidity.sh last-day-inside.txt last-day-CWQB.txt > $ROOT_WWW/last-day-humidity.svg
+    if [ -s last-day-inside.txt -o -s last-day-CWQB.txt ]; then
+	plot-simple-temperature.sh last-day-inside.txt last-day-CWQB.txt > $ROOT_WWW/last-day-temperature.svg
+	plot-simple-humidity.sh last-day-inside.txt last-day-CWQB.txt > $ROOT_WWW/last-day-humidity.svg
+    fi
 
     # graph inside temperatures over the current week
     WEEK_MIN=$(date -d 'last saturday + 1 day' +%Y%m%d)
     WEEK_MAX=$(date -d "$WEEK_MIN + 7 days" +%Y%m%d)
     extract-data.py --date-min "${WEEK_MIN}T0000" --date-max "${WEEK_MAX}T0000" < inside.txt \
         | map-to-daily-data-set.awk > last-week-inside.txt
-    plot-7days-temperature.sh last-week-inside.txt "$WEEK_MIN - $WEEK_MAX" > $ROOT_WWW/last-week-inside-temperature.svg
-    cp $ROOT_WWW/last-week-inside-temperature.svg $ROOT_WWW/archives-weekly-inside/$WEEK_MIN.svg
+    if [ -s last-week-inside.txt ]; then
+	plot-7days-temperature.sh last-week-inside.txt "$WEEK_MIN - $WEEK_MAX" > $ROOT_WWW/last-week-inside-temperature.svg
+	cp $ROOT_WWW/last-week-inside-temperature.svg $ROOT_WWW/archives-weekly-inside/$WEEK_MIN.svg
+    fi
 
     YESTERDAY=$(date -d 'yesterday' +%Y%m%d)
     if [ ! -f "$ROOT_WWW/archives/$YESTERDAY-temperature.svg" ]; then
@@ -36,4 +40,6 @@ while sleep 10m; do
         plot-simple-temperature.sh yesterday-inside.txt yesterday-CWQB.txt > $ROOT_WWW/archives/$YESTERDAY-temperature.svg
         plot-simple-humidity.sh yesterday-inside.txt yesterday-CWQB.txt > $ROOT_WWW/archives/$YESTERDAY-humidity.svg
     fi
+
+    sleep 10m
 done
